+++ projects/angular-split/src/lib/utils.ts
@@ -25,3 +25,11 @@ export function getPixelSize(elRef: ElementRef, direction: 'horizontal' | 'verti
     return elRef.nativeElement[(direction === 'horizontal') ? 'offsetWidth' : 'offsetHeight'];
 
 }
+
+export function getInputBoolean(v: any) {
+    return (typeof(v) === 'boolean') ? v : (v === 'false' ? false : true);
+}
+
+export function isValidTotalSize(total: number): boolean {
+    return total > .999 && total < 1.001;
+}