+++ src/app/component/examples/AComponent.ts
@@ -1,11 +1,13 @@
 import { ExampleEnum } from '../../data/enum';
+import { formatDate } from '../../service/utils';
 
 
 export class AComponent {
     exampleEnum = ExampleEnum
 
     testChangeDetectorRun() {
-        console.log('Change detector run just now !', Date.now())
+        console.log(`${ formatDate(new Date()) } > AComponent.ts - Change detector run just now!`);
+
         return '';
     }
 }